Outsource augmentation
