+++ kornia/augmentation/_2d/earasing.py
@@ -1,0 +1,83 @@
+from typing import Dict, Optional, Tuple, Union, cast
+
+import torch
+
+from kornia.augmentation.base_2d import IntensityAugmentationBase2D
+from kornia.geometry.bbox import bbox_generator, bbox_to_mask
+
+from .. import random_generator as rg
+
+
+class RandomErasing(IntensityAugmentationBase2D):
+    r"""Erase a random rectangle of a tensor image according to a probability p value.
+
+    .. image:: _static/img/RandomErasing.png
+
+    The operator removes image parts and fills them with zero values at a selected rectangle
+    for each of the images in the batch.
+
+    The rectangle will have an area equal to the original image area multiplied by a value uniformly
+    sampled between the range [scale[0], scale[1]) and an aspect ratio sampled
+    between [ratio[0], ratio[1])
+
+    Args:
+        p: probability that the random erasing operation will be performed.
+        scale: range of proportion of erased area against input image.
+        ratio: range of aspect ratio of erased area.
+        same_on_batch: apply the same transformation across the batch.
+        keepdim: whether to keep the output shape the same as input (True) or broadcast it
+                        to the batch form (False).
+
+    Shape:
+        - Input: :math:`(C, H, W)` or :math:`(B, C, H, W)`, Optional: :math:`(B, 3, 3)`
+        - Output: :math:`(B, C, H, W)`
+
+    Note:
+        Input tensor must be float and normalized into [0, 1] for the best differentiability support.
+        Additionally, this function accepts another transformation tensor (:math:`(B, 3, 3)`), then the
+        applied transformation will be merged int to the input transformation tensor and returned.
+
+    Examples:
+        >>> rng = torch.manual_seed(0)
+        >>> inputs = torch.ones(1, 1, 3, 3)
+        >>> aug = RandomErasing((.4, .8), (.3, 1/.3), p=0.5)
+        >>> aug(inputs)
+        tensor([[[[1., 0., 0.],
+                  [1., 0., 0.],
+                  [1., 0., 0.]]]])
+
+    To apply the exact augmenation again, you may take the advantage of the previous parameter state:
+        >>> input = torch.randn(1, 3, 32, 32)
+        >>> aug = RandomErasing((.4, .8), (.3, 1/.3), p=1.)
+        >>> (aug(input) == aug(input, params=aug._params)).all()
+        tensor(True)
+    """
+
+    # Note: Extra params, inplace=False in Torchvision.
+    def __init__(
+        self,
+        scale: Union[torch.Tensor, Tuple[float, float]] = (0.02, 0.33),
+        ratio: Union[torch.Tensor, Tuple[float, float]] = (0.3, 3.3),
+        value: float = 0.0,
+        return_transform: bool = False,
+        same_on_batch: bool = False,
+        p: float = 0.5,
+        keepdim: bool = False,
+    ) -> None:
+        super().__init__(p=p, return_transform=return_transform, same_on_batch=same_on_batch, keepdim=keepdim)
+        self.scale = scale
+        self.ratio = ratio
+        self.value: float = float(value)
+        self._param_generator = cast(rg.RectangleEraseGenerator, rg.RectangleEraseGenerator(scale, ratio, float(value)))
+
+    def apply_transform(
+        self, input: torch.Tensor, params: Dict[str, torch.Tensor], transform: Optional[torch.Tensor] = None
+    ) -> torch.Tensor:
+        _, c, h, w = input.size()
+        values = params["values"].unsqueeze(-1).unsqueeze(-1).unsqueeze(-1).repeat(1, *input.shape[1:]).to(input)
+
+        bboxes = bbox_generator(params["xs"], params["ys"], params["widths"], params["heights"])
+        mask = bbox_to_mask(bboxes, w, h)  # Returns B, H, W
+        mask = mask.unsqueeze(1).repeat(1, c, 1, 1).to(input)  # Transform to B, c, H, W
+        transformed = torch.where(mask == 1.0, values, input)
+        return transformed
