+++ kornia/augmentation/_2d/gaussian_noise.py
@@ -1,0 +1,57 @@
+from typing import Dict, Optional
+
+import torch
+
+from kornia.augmentation.base_2d import IntensityAugmentationBase2D
+
+
+class RandomGaussianNoise(IntensityAugmentationBase2D):
+    r"""Add gaussian noise to a batch of multi-dimensional images.
+
+    .. image:: _static/img/RandomGaussianNoise.png
+
+    Args:
+        mean: The mean of the gaussian distribution.
+        std: The standard deviation of the gaussian distribution.
+        return_transform: if ``True`` return the matrix describing the transformation applied to each
+            input tensor. If ``False`` and the input is a tuple the applied transformation won't be concatenated.
+        same_on_batch: apply the same transformation across the batch.
+        p: probability of applying the transformation.
+        keepdim: whether to keep the output shape the same as input (True) or broadcast it
+                 to the batch form (False).
+    Examples:
+        >>> rng = torch.manual_seed(0)
+        >>> img = torch.ones(1, 1, 2, 2)
+        >>> RandomGaussianNoise(mean=0., std=1., p=1.)(img)
+        tensor([[[[ 2.5410,  0.7066],
+                  [-1.1788,  1.5684]]]])
+
+    To apply the exact augmenation again, you may take the advantage of the previous parameter state:
+        >>> input = torch.randn(1, 3, 32, 32)
+        >>> aug = RandomGaussianNoise(mean=0., std=1., p=1.)
+        >>> (aug(input) == aug(input, params=aug._params)).all()
+        tensor(True)
+    """
+
+    def __init__(
+        self,
+        mean: float = 0.0,
+        std: float = 1.0,
+        return_transform: bool = False,
+        same_on_batch: bool = False,
+        p: float = 0.5,
+        keepdim: bool = False,
+    ) -> None:
+        super().__init__(
+            p=p, return_transform=return_transform, same_on_batch=same_on_batch, p_batch=1.0, keepdim=keepdim
+        )
+        self.flags = dict(mean=mean, std=std)
+
+    def generate_parameters(self, shape: torch.Size) -> Dict[str, torch.Tensor]:
+        noise = torch.randn(shape)
+        return dict(noise=noise)
+
+    def apply_transform(
+        self, input: torch.Tensor, params: Dict[str, torch.Tensor], transform: Optional[torch.Tensor] = None
+    ) -> torch.Tensor:
+        return input + params["noise"].to(input.device) * self.flags["std"] + self.flags["mean"]
