+++ kornia/augmentation/_2d/grayscale.py
@@ -1,0 +1,65 @@
+from typing import Dict, Optional
+
+import torch
+
+from kornia.augmentation.base_2d import IntensityAugmentationBase2D
+from kornia.color import rgb_to_grayscale
+
+
+class RandomGrayscale(IntensityAugmentationBase2D):
+    r"""Apply random transformation to Grayscale according to a probability p value.
+
+    .. image:: _static/img/RandomGrayscale.png
+
+    Args:
+        p: probability of the image to be transformed to grayscale.
+        return_transform: if ``True`` return the matrix describing the transformation applied to each
+                          input tensor. If ``False`` and the input is a tuple the applied transformation
+                          won't be concatenated.
+        same_on_batch: apply the same transformation across the batch.
+        keepdim: whether to keep the output shape the same as input (True) or broadcast it
+                        to the batch form (False).
+
+    Shape:
+        - Input: :math:`(C, H, W)` or :math:`(B, C, H, W)`, Optional: :math:`(B, 3, 3)`
+        - Output: :math:`(B, C, H, W)`
+
+    .. note::
+        This function internally uses :func:`kornia.color.rgb_to_grayscale`.
+
+    Examples:
+        >>> rng = torch.manual_seed(0)
+        >>> inputs = torch.randn((1, 3, 3, 3))
+        >>> aug = RandomGrayscale(p=1.0)
+        >>> aug(inputs)
+        tensor([[[[-1.1344, -0.1330,  0.1517],
+                  [-0.0791,  0.6711, -0.1413],
+                  [-0.1717, -0.9023,  0.0819]],
+        <BLANKLINE>
+                 [[-1.1344, -0.1330,  0.1517],
+                  [-0.0791,  0.6711, -0.1413],
+                  [-0.1717, -0.9023,  0.0819]],
+        <BLANKLINE>
+                 [[-1.1344, -0.1330,  0.1517],
+                  [-0.0791,  0.6711, -0.1413],
+                  [-0.1717, -0.9023,  0.0819]]]])
+
+    To apply the exact augmenation again, you may take the advantage of the previous parameter state:
+        >>> input = torch.randn(1, 3, 32, 32)
+        >>> aug = RandomGrayscale(p=1.0)
+        >>> (aug(input) == aug(input, params=aug._params)).all()
+        tensor(True)
+    """
+
+    def __init__(
+        self, return_transform: bool = False, same_on_batch: bool = False, p: float = 0.1, keepdim: bool = False
+    ) -> None:
+        super().__init__(p=p, return_transform=return_transform, same_on_batch=same_on_batch, keepdim=keepdim)
+
+    def apply_transform(
+        self, input: torch.Tensor, params: Dict[str, torch.Tensor], transform: Optional[torch.Tensor] = None
+    ) -> torch.Tensor:
+        # Make sure it returns (*, 3, H, W)
+        grayscale = torch.ones_like(input)
+        grayscale[:] = rgb_to_grayscale(input)
+        return grayscale
