+++ kornia/augmentation/_2d/invert.py
@@ -1,0 +1,60 @@
+from typing import Dict, Optional, Union
+
+import torch
+
+from kornia.augmentation.base_2d import IntensityAugmentationBase2D
+from kornia.enhance import invert
+
+
+class RandomInvert(IntensityAugmentationBase2D):
+    r"""Invert the tensor images values randomly.
+
+    .. image:: _static/img/RandomInvert.png
+
+    Args:
+        max_val: The expected maximum value in the input tensor. The shape has to
+          according to the input tensor shape, or at least has to work with broadcasting.
+        return_transform: if ``True`` return the matrix describing the transformation applied to each
+            input tensor. If ``False`` and the input is a tuple the applied transformation won't be concatenated.
+        same_on_batch: apply the same transformation across the batch.
+        p: probability of applying the transformation.
+        keepdim: whether to keep the output shape the same as input (True) or broadcast it
+                 to the batch form (False).
+    .. note::
+        This function internally uses :func:`kornia.enhance.invert`.
+
+    Examples:
+        >>> rng = torch.manual_seed(0)
+        >>> img = torch.rand(1, 1, 5, 5)
+        >>> inv = RandomInvert()
+        >>> inv(img)
+        tensor([[[[0.4963, 0.7682, 0.0885, 0.1320, 0.3074],
+                  [0.6341, 0.4901, 0.8964, 0.4556, 0.6323],
+                  [0.3489, 0.4017, 0.0223, 0.1689, 0.2939],
+                  [0.5185, 0.6977, 0.8000, 0.1610, 0.2823],
+                  [0.6816, 0.9152, 0.3971, 0.8742, 0.4194]]]])
+
+    To apply the exact augmenation again, you may take the advantage of the previous parameter state:
+        >>> input = torch.randn(1, 3, 32, 32)
+        >>> aug = RandomInvert(p=1.)
+        >>> (aug(input) == aug(input, params=aug._params)).all()
+        tensor(True)
+    """
+
+    def __init__(
+        self,
+        max_val: Union[float, torch.Tensor] = torch.tensor(1.0),
+        return_transform: bool = False,
+        same_on_batch: bool = False,
+        p: float = 0.5,
+        keepdim: bool = False,
+    ) -> None:
+        super().__init__(
+            p=p, return_transform=return_transform, same_on_batch=same_on_batch, p_batch=1.0, keepdim=keepdim
+        )
+        self.flags = dict(max_val=max_val)
+
+    def apply_transform(
+        self, input: torch.Tensor, params: Dict[str, torch.Tensor], transform: Optional[torch.Tensor] = None
+    ) -> torch.Tensor:
+        return invert(input, torch.as_tensor(self.flags["max_val"], device=input.device, dtype=input.dtype))
