+++ kornia/augmentation/_3d/__init__.py
@@ -1,0 +1,8 @@
+from .affine import RandomAffine3D
+from .center_crop import CenterCrop3D
+from .crop import RandomCrop3D
+from .equalize import RandomEqualize3D
+from .flip import RandomDepthicalFlip3D, RandomHorizontalFlip3D, RandomVerticalFlip3D
+from .motion_blur import RandomMotionBlur3D
+from .perspective import RandomPerspective3D
+from .rotation import RandomRotation3D
