+++ kornia/augmentation/_3d/rotation.py
@@ -1,0 +1,116 @@
+from typing import Dict, Optional, Tuple, Union, cast
+
+import torch
+
+from kornia.constants import Resample
+from kornia.geometry import affine3d
+from kornia.geometry.transform.affwarp import _compute_rotation_matrix3d, _compute_tensor_center3d
+
+from .. import random_generator as rg
+from ..base_3d import AugmentationBase3D
+
+
+class RandomRotation3D(AugmentationBase3D):
+    r"""Apply random rotations to 3D volumes (5D tensor).
+
+    Input should be a tensor of shape (C, D, H, W) or a batch of tensors :math:`(B, C, D, H, W)`.
+    If Input is a tuple it is assumed that the first element contains the aforementioned tensors and the second,
+    the corresponding transformation matrix that has been applied to them. In this case the module
+    will rotate the tensors and concatenate the corresponding transformation matrix to the
+    previous one. This is especially useful when using this functionality as part of an ``nn.Sequential`` module.
+
+    Args:
+        degrees: Range of degrees to select from.
+            If degrees is a number, then yaw, pitch, roll will be generated from the range of (-degrees, +degrees).
+            If degrees is a tuple of (min, max), then yaw, pitch, roll will be generated from the range of (min, max).
+            If degrees is a list of floats [a, b, c], then yaw, pitch, roll will be generated from (-a, a), (-b, b)
+            and (-c, c).
+            If degrees is a list of tuple ((a, b), (m, n), (x, y)), then yaw, pitch, roll will be generated from
+            (a, b), (m, n) and (x, y).
+            Set to 0 to deactivate rotations.
+        resample:
+        return_transform: if ``True`` return the matrix describing the transformation applied to each
+          input tensor. If ``False`` and the input is a tuple the applied transformation won't be concatenated.
+        same_on_batch: apply the same transformation across the batch.
+        align_corners: interpolation flag.
+        keepdim: whether to keep the output shape the same as input (True) or broadcast it
+          to the batch form (False).
+
+    Shape:
+        - Input: :math:`(C, D, H, W)` or :math:`(B, C, D, H, W)`, Optional: :math:`(B, 4, 4)`
+        - Output: :math:`(B, C, D, H, W)`
+
+    Note:
+        Input tensor must be float and normalized into [0, 1] for the best differentiability support.
+        Additionally, this function accepts another transformation tensor (:math:`(B, 4, 4)`), then the
+        applied transformation will be merged int to the input transformation tensor and returned.
+
+    Examples:
+        >>> rng = torch.manual_seed(0)
+        >>> input = torch.rand(1, 1, 3, 3, 3)
+        >>> aug = RandomRotation3D((15., 20., 20.), p=1.0, return_transform=True)
+        >>> aug(input)
+        (tensor([[[[[0.3819, 0.4886, 0.2111],
+                   [0.1196, 0.3833, 0.4722],
+                   [0.3432, 0.5951, 0.4223]],
+        <BLANKLINE>
+                  [[0.5553, 0.4374, 0.2780],
+                   [0.2423, 0.1689, 0.4009],
+                   [0.4516, 0.6376, 0.7327]],
+        <BLANKLINE>
+                  [[0.1605, 0.3112, 0.3673],
+                   [0.4931, 0.4620, 0.5700],
+                   [0.3505, 0.4685, 0.8092]]]]]), tensor([[[ 0.9722,  0.1131, -0.2049,  0.1196],
+                 [-0.0603,  0.9669,  0.2478, -0.1545],
+                 [ 0.2262, -0.2286,  0.9469,  0.0556],
+                 [ 0.0000,  0.0000,  0.0000,  1.0000]]]))
+
+    To apply the exact augmenation again, you may take the advantage of the previous parameter state:
+        >>> input = torch.rand(1, 3, 32, 32, 32)
+        >>> aug = RandomRotation3D((15., 20., 20.), p=1.)
+        >>> (aug(input) == aug(input, params=aug._params)).all()
+        tensor(True)
+    """
+
+    def __init__(
+        self,
+        degrees: Union[
+            torch.Tensor,
+            float,
+            Tuple[float, float, float],
+            Tuple[Tuple[float, float], Tuple[float, float], Tuple[float, float]],
+        ],
+        resample: Union[str, int, Resample] = Resample.BILINEAR.name,
+        return_transform: bool = False,
+        same_on_batch: bool = False,
+        align_corners: bool = False,
+        p: float = 0.5,
+        keepdim: bool = False,
+    ) -> None:
+        super().__init__(p=p, return_transform=return_transform, same_on_batch=same_on_batch, keepdim=keepdim)
+        self.flags = dict(resample=Resample.get(resample), align_corners=align_corners)
+        self._param_generator = cast(rg.RotationGenerator3D, rg.RotationGenerator3D(degrees))
+
+    def compute_transformation(self, input: torch.Tensor, params: Dict[str, torch.Tensor]) -> torch.Tensor:
+        yaw: torch.Tensor = params["yaw"].to(input)
+        pitch: torch.Tensor = params["pitch"].to(input)
+        roll: torch.Tensor = params["roll"].to(input)
+
+        center: torch.Tensor = _compute_tensor_center3d(input)
+        rotation_mat: torch.Tensor = _compute_rotation_matrix3d(yaw, pitch, roll, center.expand(yaw.shape[0], -1))
+
+        # rotation_mat is B x 3 x 4 and we need a B x 4 x 4 matrix
+        trans_mat: torch.Tensor = torch.eye(4, device=input.device, dtype=input.dtype).repeat(input.shape[0], 1, 1)
+        trans_mat[:, 0] = rotation_mat[:, 0]
+        trans_mat[:, 1] = rotation_mat[:, 1]
+        trans_mat[:, 2] = rotation_mat[:, 2]
+
+        return trans_mat
+
+    def apply_transform(
+        self, input: torch.Tensor, params: Dict[str, torch.Tensor], transform: Optional[torch.Tensor] = None
+    ) -> torch.Tensor:
+        transform = cast(torch.Tensor, transform)
+        return affine3d(
+            input, transform[..., :3, :4], self.flags["resample"].name.lower(), "zeros", self.flags["align_corners"]
+        )
