+++ test/augmentation/test_augmentation.py
@@ -30,10 +30,10 @@ from kornia.augmentation import (
     RandomThinPlateSpline,
     RandomVerticalFlip,
 )
-from kornia.augmentation.base import AugmentationBase2D
-from kornia.constants import pi, Resample
+from kornia.augmentation.base_2d import AugmentationBase2D
+from kornia.constants import Resample, pi
 from kornia.geometry import transform_points
-from kornia.testing import assert_close, BaseTester, default_with_one_parameter_changed
+from kornia.testing import BaseTester, assert_close, default_with_one_parameter_changed
 from kornia.utils import create_meshgrid
 from kornia.utils.helpers import _torch_inverse_cast
